Update Jest tests for SearchPageHelper and the refactored SearchModule.
